More granular proof dependencies (#4851)
# Description

This PR refactors the proof dependency data types and the warning
generation code to do several things:

* Distinguish between the various types of assumption that can show up.
We currently report redundancies only for assumptions that come directly
from assume and assert statements and for requires clauses. Other types
are difficult for the user to control, but will likely be useful to
treat separately in the future.
* Distinguish between the split parts of `requires` and `ensures`
clauses.
* Clarify that unnecessary assumptions and assertions may be only partly
unnecessary. A next step will be to indicate which part is unnecessary,
but that turns out to require changing verification conditions in a way
that may affect performance and brittleness, so I'm going to leave that
for a separate PR that I hope to land after we release 4.4.

Fixes #4719 

# How has this been tested?
This changes the expected output of a few existing tests.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.