fix: Fix new resolution of unary minus (#4737)
Previously, an expression `-d` could be typed as `nat` given `d: nat`.
This caused a verification error, trying to prove that the result is a
`nat`. This PR fixes this problem in the new resolver.

Note, the old resolver still has this problem, but it's more difficult
to fix there. Thus, this PR solves the problem just for the new resolver
(and I propose we ignore the problem in the old resolver, since we'll be
switching everything over to the new resolver in the not too distant
future).


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>