Fix: Formatting on the command line works again (#4950)
This PR makes sure StreamReaders are closed after use, which prevented
formatting from overwriting the files randomly. The reason why to do
this despite an obvious memory increase is that in [all places where we
use this
StreamReader](https://github.com/dafny-lang/dafny/pull/4918#discussion_r1444530522),
we actually already read to the end and don't make use of the streaming.

Before, the "dafny format" command was throwing exceptions on large
files, and now it no longer throws an exception (test added for that)

Even better, this solved the issue that when I create a blank Dafny file
and try to save it, it used to complain with "Failed to save
'todelete.dfy': Unable to write file". (Fixes
https://github.com/dafny-lang/ide-vscode/issues/452)

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>