Implement `:resource_limit` attribute (#4975)
### Description

This is an alternative to `:rlimit` that doesn't multiply by 1000 before
sending the value to Z3. With the addition of this new attribute,
`:rlimit` is deprecated.

The `--resource-limit` flag now also passes through the value unchanged,
without a multiplication. This flag is currently not used widely, so the
impact of this breaking change is minimal.

The `:resource_limit` attribute and `--resource-limit` flag now accept
`K`, `M`, and `G` suffixes to allow shorter, easier-to-read limits.

### How has this been tested?

The integration tests have been updated to use `:resource_limit`
wherever they used `:rlimit` previously. Some `--resource-limit` and
`:resource_limit` uses have been updated to use suffixes.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.