fix: Fix parsing error and a crash with non-reference-trait base type (#4956)
This PR fixes two bugs.

One bug is to insist that a subset type name its constrained value (that
is, the `x:` is not allowed to be omitted, unlike in a `newtype`
declaration). This had caused a crash in the resolver (different crashes
in the legacy and new resolver, it turns out). The crash had been
reported in https://github.com/dafny-lang/dafny/issues/4946.

The other bug is that, even with a given identifier name `x:`, the
verifier would crash when trying to construct a default witness (if the
base type is a non-reference trait and there is no explicit `witness`
clause).

Fixes https://github.com/dafny-lang/dafny/issues/4946


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>