Refactor tests to avoid issues with skipped errors
Boogie is not always guaranteed to report all errors in a procedure,
and often doesn’t report a later error that follows a previous on a
straight line path. It essentially determines that the later one is
dead code and skips it.

Refactoring one test allows us to ensure we get all of the errors we
expect to get. For the other, it’s safe to just skip that error.