Use resource limit in most places in test suite (#4972)
### Description

A few exceptions remain, but this will allow testing the general idea. There is still _also_ a time limit, because that does help ensure that the test suite terminates in a predictable amount of time. Over time, it may make sense to reduce the overall resource limit and raise it locally for the few tests that need it.

Fixes #2985

### How has this been tested?

It changes the behavior of the existing test suite. Passing existing tests confirms that it's working correctly.

By submitting this pull request, I confirm that my contribution is made under the terms of the MIT license.