Make smaller changes to compilers (#5052)
This PR contains some minor changes to the compilers. These changes have
been carved off from a larger set of changes (for supporting general
`newtype`s and `is` expressions), to simplify the reviewing of the
latter.

The changes in this PR are:
* Rename parameter of `EmitDowncastVariableAssignment`
(`collectionElementType` to `sourceType`).
* Add `nativeType` parameter to ` EmitBitvectorTruncation` (to allow it
to be different than `bvType.NativeType`).
* In some places, don't assume that native type of bitvector type is
always as wide as the bitvector. (This happens to be true today, but it
may not be true in the future (i) when a `newtype` can have a bitvector
base type, which will happen soon, and (ii) if `IntBoundedPool` starts
supporting bitvector types.

### How has this been tested?

Since the changes are meant to preserve current behaviors (but with an
eye to the up-and-coming PRs with larger changes), this PR does not make
any changes to tests.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>