fix: Fill in type arguments in implicit function-by-method postcondition (#5068)
This PR fills in any type arguments `X` to the `result = F<X>(args)`
postcondition that's generated for the method part of a
`function-by-method` declaration.

Fixes #4998 

Reviewer notes: The desugaring of `function-by-method` is done in two
places in the code. I filled in the type arguments in both places.
However, in the second place (which is for `{:test}` functions/methods),
`dafny` would crash if any type parameters were declared (even for type
parameters that were not auto-init `(0)`). Since the `{:test}` was
already not allowed for functions/methods with parameters, I also added
error messages if `{:test}` is used with a function/method with type
parameters.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>