Filter assertions (#5031)
### Description
- Allow using `--filter-position` to only verify individual assertions.

The implementation relies on a new version of Boogie, which changes the
API so that Boogie returns not a list of implementations but instead of
a list of 'verification tasks', which map to individual SMT executions
or 'assertion batches'. When using `--isolate-assertions`, each
assertion is given such as task, and it becomes easy to filter these
tasks/assertions based on a file position.

#### Background of some of the code changes
This change in the Boogie API means Dafny is not longer updated when all
the tasks for a particular Boogie implementation are finished, which
effects a few things that rely on all tasks in a Boogie implementation
to be done:
 - Computation of redundant assertions
- Gutter icons (maybe it conceptually does not depend on Boogie
implementations, I'm not sure, but the current implementation cares
about Boogie implementations)

There is no change in behavior for the above, but the code had to be
updated so on the Dafny side we track whether everything for a
particular implementation is done.

### How has this been tested?
- Added new cases to the existing littish test `filter.dfy`
- Some gutter icons tests have been updated, which is I think due to an
improvement. The last batch completed + implementation completed steps
are now combined, so two updates that would always immediately follow
each other are now squashed together.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>