Rework warn-as-error option (#4971)
The existing option `--warn-as-error` causes everything that would
otherwise be a warning, to be recorded as an error. Because the Dafny
compiler often uses the presence of errors to determine whether to
continue, using `--warn-as-error` causes compilation to potentially halt
earlier when it's turned on. This can hide errors until warnings that
occur earlier in the pipeline have been resolved. In addition, using
this option makes it impossible for the user to see what was originally
a warning instead of an error.

### Changes
- Replace the option `--warn-as-error` with the option
`--succeed-on-warnings`. Besides having the opposite meaning, this
option no longer turns warnings into errors, nor does it reduce the
amount of diagnostics collected in compilation. When turned off, the
default, any warnings will cause the CLI to return an error exit code
and prevent the execution of the Dafny code. All forms of execution,
whether directly executing it or through the creation of build artifacts
such as generated code (cs/java/go/etc) or further compiled generated
code (dll/jar/etc), are prevented.
- A doo file that was created using `--succeed-on-warnings` causes a
_warning_ if used by a consumer that does not use it.
- Change the behavior of `--warn-deprecation=false` so that deprecation
diagnostics are reported as information, which means they show up in the
IDE
- Flip the behavior of `--warn-deprecation` and change the name to
`--allow-deprecation`, so the default is now false, which is common for
boolean options. It's slightly nicer for user to type
`--allow-deprecation` instead of `--warn-deprecation=false`.

### Caveats
- When a user is developing and wants to quickly run their code to test
something, but the code has warnings, the user must add
`--succeed-on-warnings` to their `dafny run` invocation.
- This change is not backwards compatible. 
  - Dafny users must remove any usages of `--warn-as-error`
- If users have builds that were passing with warnings, they have to add
`--succeed-on-warnings` to allow them to still pass
- If users upgrade to a new Dafny version, and are not using
`--succeed-on-warnings`, and do not want to migrate off of deprecated
feature, they will have to use `--allow-deprecation`. In the future we
will introduce `dafny migrate` to enable use to automatically move off
of deprecated features, allowing them to instantly upgrade to newer
Dafny versions without having to use `--allow-deprecation` or
`--succeed-on-warnings`.
- Any instance of `--warn-deprecation false` must be replaced with
`--allow-deprecation`

### How has this been tested?
- Updated tests

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Mikaël Mayer <MikaelMayer@users.noreply.github.com>