Use async main methods to avoid waiting for tasks at the top level (#4585)
### Changes
Remove three occurrences of `#pragma warning disable VSTHRD002` by
relying on async Main methods. An effect of this change might be that
Dafny will use fewer threads. I was triggered to make this change when I
was debugging Dafny and noticed some threads always waiting at the
locations where this PR removes the Wait.

### Testing
This is a refactoring which is covered by existing tests.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>