feat: Add general `is` support, and allow newtypes for bool/char/bv (#5061)
### Overview

This PR provides two new features:
* Allow a full repertoire of `is` tests in the language, and support
such tests at run time when possible. (Previously, `is` was supported in
a rather spotty way.)
* Allow `newtype`s based on any primitive type. (Previously, newtypes
could only be based on `int` and `real`. Now, they can also be based on
`bool`, `char`, and bitvector types.) Bitvector-based newtypes allow
`:nativeType` annotations (but do not analyze value ranges, as is done
for `int`).

The PR is big. It has been made somewhat smaller by dividing up the
changes into some smaller PRs that have been merged. There are currently
two outstanding such PRs (#5051 and #5053), but (as the initial comments
on those PRs mention) they may not be as useful to review in isolation.
Therefore, I'm submitting all the code in this PR (which includes the
changes in (#5051 and #5053). So, those other ones can be ignored or can
be reviewed/accepted before the present PR.

The PR contains two features, one related to `is` and one related to
`newtype`. I had tried to tease these apart when I was trying to chop
everything up into smaller PRs. But that worked out to be trickier than
I had thought, and the tests for the two features are somewhat mixed
into common files, too. So, I'm submitting this PR and won't try to chop
it any further, unless that becomes really necessary for reviewing.

While writing the code for these features, various bugs were discovered
and fixed. This PR includes regression tests for those bugs; however,
these tests may or may not be labeled as such, and there may or may not
have been an associated Issue for every bug fixed.

### New rules for `is` type tests

An expression `e as T` or `e is T` is now permitted in the following
situations, where `F` (for "from") is the static type of expression `e`:

    T ::> F    or    F ::> T

where `X ::> Y` means

* `X'` is a supertype of `Y'`, or
* `X'` and `Y'` are various bitvector types, or
* `X'` is `int` and `Y'` is one of `int`, `char`, a bitvector type,
`ORDINAL`, or `real`, or
* (*) `X'` is `real` and `Y'` is one of `int`, `char`, a bitvector type,
or `ORDINAL`, or
* (*) `X'` is `char` and `Y'` is one of `int`, a bitvector type, or
`ORDINAL`, or
* (*) `X'` is a bitvector type and `Y'` is `int`, `char`, `ORDINAL`, or
`real`,

where `X'` and `Y'` are the `newtype` ancestors of `X` and `Y`,
respectively.

Several of these cases are new, including many cases when `T` is a
subset type or a newtype, or when `F` and `T` are equal. This PR thus
extends `as` and `is` to cases that one would just expect to be there.

In this definition, the (*) cases look overly permissive to me. They are
included here, because Dafny accepts those today. However, the matrix of
possibilities of various conversions is already highly complex, so I am
interested in simplifying the language to throw out the cases that are
not likely to be used and can easily be done as two conversions via
`int`. Therefore, the (*) cases are allowed only with
`--general-newtypes=false` (which is the still the default). This PR
omits the (*) cases under `--general-newtypes`.

What was just said applies to `as` and `is` expressions. In addition,
there are restrictions on which `is` expressions can be compiled, and
those restrictions also apply to implicit type tests that are part of
comprehensions. (Comprehensions already supported some forms of
subset-type constraint tests. This PR extends those, also applies them
to newtypes, and restructures how/when the resolvers figures out which
constraints are compilable.) The additional restrictions for `is` (note,
not `as`) to be compilable are:

* When the `is` expression is allowed because of the subtype case,
suppose the ancestor of `F` is `C<...>` and the ancestor of `T` is
`D<...>`. Then, to be compilable, the type parameters of `D` must be
injective in the type parameters of `C`. For example, the test is
compilable given `class D<K> extends C<K>` or `datatype D<K> extends
C<seq<K>>`, and the test is not compilable if `datatype D<K> extends
D<int>` or if `D` is a class with type parameters and `C` is `object`.
* If `T` is a subset type or newtype that involves constraints, then
either `T` equals `F` or those constraints must be compilable. (This
could be extended in the future to say that the constraints _between_
the types `T` and `F` must be compilable.)

### New `newtype`s

Under the option `--general-newtypes`, this PR allows a `newtype` to use
`bool`, `char`, or a bitvector type as a base type.

Note that using `--general-newtypes` causes the (*) cases above to be
omitted. So, to opt into the new `newtype`s, one also opts out of the
arcane legacy `is` expressions.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>