Translate only recognized attributes (#5107)
### Description

Fixes #4204

Previously, most attributes were copied from Dafny to Boogie without any
filtering. This could result in unsoundness, as described in issue
#4204. This PR changes the translator to recognize a fixed set of
attributes, and to translate or copy each one as appropriate. It checks
types, as well, so, for example, `{:timeLimit N}` will only be
translated if `N` is an integer.

This should have no user-visible changes except for code (if it exists)
that attempts to use Boogie attributes that don't make sense for Dafny.

### How has this been tested?

The existing test suite should cover all the relevant functionality.

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.