chore: Rename EmitIsRuneTest to EmitIsUnicodeScalarValueTest (#5115)
This PR renames a virtual method in the compiler. The change was
suggested in the (now abandoned and closed) PR
https://github.com/dafny-lang/dafny/pull/5053.

Note, the PR does not rename other occurrences of "rune". In particular,
.NET has a class `System.Text.Rune` and the Dafny runtime (therefore?)
has a class named `Rune` (and inside it, a method `IsRune`). These have
not been renamed.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>