fix: Use reveal_ constant as function argument in override axiom (#5111)
Fixes #5017

Let verifier understand opaque function overrides, supporting both when
the overridden function is opaque and non-opaque. Revealing such a
function for one overriding type has the effect of revealing it for all
overriding types.

Also, forbid the case where a function is opaque in a parent trait and
the function override is not opaque. Previously, this had caused a
crash.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>