chore: Update general-traits error message (#5191)
Previously, the error message for using `extends` with a non-reference
type said that this was a beta feature. It is supported, but requires
the CLI option. This PR rewords these error messages.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>