set RangeToken for ParensExpression (#5219)
### Description
RangeToken is not set for ParensExpression in
https://github.com/dafny-lang/dafny/blob/99ad60af5b23ad83b2e49768cb32c8d2599cd60d/Source/DafnyCore/Dafny.atg#L3648
and it is assumed that it is going to be set in ParensExpression
definition itself. However, RangeToken is not set in ParensExpression
and this PR fixes it.

### How has this been tested?
Let me know if it is required to add a test for this. I couldn't find a
similar test on RangeTokens to replicate.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>