Do not use DafnyConsolePrinter when using the modular CLI front-end (#5214)
### Description
- `DafnyConsolePrinter` globally stores information resulting from
verification, which consumes memory. Previously it was causing out of
memory issues, and then we reduced the amount of memory it consumed, but
it still consumes an amount of memory linear to the amount of
verification you're doing in the CLI call. By not using this class,
which is not required when using the modular CLI front-end, we remove
one cause of memory leaks.

### How has this been tested?
- Refactoring, covered by existing tests.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>