Support Counterexamples with New CLI Interface  (#5135)
### Description
After the recent changes to the new CLI, Dafny no longer displays
counterexamples when requested. This PR fixes the issue and also updates
the tests to the new CLI to avoid this problem in the future.

As part of this PR, I also moved counterexample code to `DafnyCore`.
This seems to be the best way to support the changes to the new CLI, so
that files in `DafnyCore/Pipeline` can process counterexamples. It also
makes more sense to have counterexample there rather than in
`DafnyLanguageServer`, I think.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Aleksandr Fedchin <fedchina@amazon.com>
Co-authored-by: Remy Willems <rwillems@amazon.com>